- Fix Transform's append and prepend methods. They have been used in the complete opposite. - Fix UI transform hierarchy. - WIP: Show CPU cycle counters in the debug overlay.

+++ projects/PathosEngine/src/pathos/render/render_overlay.h
@@ -38,7 +38,7 @@ namespace pathos {
 		DisplayObject2DProxy* root = nullptr;
 
 	private:
-		void calculateTransformNDC(uint16 windowWidth, uint16 windowHeight);
+		void calculateTransformNDC(uint32 windowWidth, uint32 windowHeight);
 
 		Transform toNDC; // screen coordinates to NDC
